requirements.txt
openai==1.3.5
azure-identity==1.15.0
python-dotenv==1.0.0
pandas==2.1.3